+++ dookie/visualize.hpp
@@ -7,7 +7,7 @@
 struct Context;
 struct Visualizer
 {
-  void Visualize(const std::vector<uint16_t> & samps, const std::vector<double> &freqs, DisplayContext & display);
+  void Visualize(const std::vector<uint8_t> & samps, const std::vector<double> &freqs, DisplayContext & display);
   void BeatStart();
   double R = 0.0;
 };
